+++ BraintreeDropIn/Models/BTDropInRequest.m
@@ -18,13 +18,18 @@
 
 - (id)copyWithZone:(__unused NSZone *)zone {
     BTDropInRequest *request = [BTDropInRequest new];
+
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated-declarations"
     request.amount = self.amount;
+#pragma clang diagnostic pop
     request.payPalRequest = self.payPalRequest;
     request.applePayDisabled = self.applePayDisabled;
     request.paypalDisabled = self.paypalDisabled;
     request.venmoDisabled = self.venmoDisabled;
     request.cardDisabled = self.cardDisabled;
     request.threeDSecureVerification = self.threeDSecureVerification;
+    request.threeDSecureRequest = self.threeDSecureRequest;
     request.cardholderNameSetting = self.cardholderNameSetting;
     request.shouldMaskSecurityCode = self.shouldMaskSecurityCode;
     request.vaultManager = self.vaultManager;