+++ DropInDemo/Demo Base/Settings/BraintreeDemoSettings.h
@@ -1,9 +1,11 @@
 #import <Foundation/Foundation.h>
 #import <BraintreeDropIn/BTDropInRequest.h>
+#import <Braintree/BraintreePaymentFlow.h>
 
 extern NSString *BraintreeDemoSettingsEnvironmentDefaultsKey;
 extern NSString *BraintreeDemoSettingsCustomEnvironmentURLDefaultsKey;
 extern NSString *BraintreeDemoSettingsThreeDSecureRequiredDefaultsKey;
+extern NSString *BraintreeDemoSettingsThreeDSecureVersionDefaultsKey;
 
 typedef NS_ENUM(NSInteger, BraintreeDemoTransactionServiceEnvironment) {
     BraintreeDemoTransactionServiceEnvironmentSandboxBraintreeSampleMerchant = 0,
@@ -25,6 +27,7 @@ typedef NS_ENUM(NSInteger, BraintreeDemoTransactionServiceThreeDSecureRequiredSt
 + (NSString *)authorizationOverride;
 + (BOOL)useTokenizationKey;
 + (BraintreeDemoTransactionServiceThreeDSecureRequiredStatus)threeDSecureRequiredStatus;
++ (BTThreeDSecureVersion)threeDSecureRequestedVersion;
 + (BOOL)useModalPresentation;
 + (BOOL)customerPresent;
 + (NSString *)customerIdentifier;